+++ src/main/java/com/mass3d/support/hibernate/HibernateGenericStore.java
@@ -375,7 +375,7 @@ public class HibernateGenericStore<T>
   }
 
   @Override
-  public T get( int id )
+  public T get( Long id )
   {
     T object = getSession().get( getClazz(), id );
